# server/docs/swagger.yaml
definitions:
  controller.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      nick_name:
        type: string
      password:
        type: string
      remark:
        type: string
      updated_at:
        type: string
      userRole:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        description: 一个操作员只能有一个角色
      username:
        type: string
    type: object
  model.UserRole:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_code:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  requests.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  response.Auth:
    properties:
      id:
        type: integer
      token:
        type: string
      user:
        type: string
    type: object
...
info:
  contact: {}
  description: golang 定时任务调度服务系统v1.0
  title: 定时任务调度服务系统v1.0
  version: V1.1
paths:
...
  /api/v1/admin/updatepassword:
    post:
      description: 更新用户密码
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户密码
      tags:
      - 用户
  /api/v1/admin/updateuserinfo:
    post:
      description: 更新用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - 用户
  /api/v1/admin/upload:
    post:
      description: 上传文件
      parameters:
      - description: 用户名、密码
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/response.Auth'
              type: object
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 上传文件
      tags:
      - 基础接口
  /api/v1/admin/userInfo:
    get:
      description: 获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户信息
      tags:
      - 用户
...
  /api/v1/menu/catalog:
    get:
      description: 获取一级菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取一级菜单列表
      tags:
      - 菜单
  /api/v1/menu/create:
    post:
      description: 创建菜单
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建菜单
      tags:
      - 菜单
  /api/v1/menu/delete:
    delete:
      description: 删除菜单
      parameters:
      - description: 菜单id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除菜单
      tags:
      - 菜单
  /api/v1/menu/list:
    get:
      description: 获取菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单
  /api/v1/menu/roleMenus/{code}:
    get:
      description: 获取菜单列表
      parameters:
      - description: 角色编码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单
  /api/v1/menu/saveRoleMenus/{code}:
    post:
      description: 保存角色菜单
      parameters:
      - description: 角色编码
        in: path
        name: code
        required: true
        type: string
      - description: 菜单id列表
        in: body
        name: menus
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 保存角色菜单
      tags:
      - 菜单
  /api/v1/menu/update:
    put:
      description: 修改菜单
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改菜单
      tags:
      - 菜单
...
  /api/v1/user/create:
    post:
      description: 创建用户
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建用户
      tags:
      - 用户
  /api/v1/user/list:
    get:
      description: 获取用户列表
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - 用户
  /api/v1/user/update:
    delete:
      description: 删除用户
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
    put:
      description: 更新用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 'code: 200 成功'
          schema:
            $ref: '#/definitions/controller.Response'
        "500":
          description: 错误返回内容
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"